requirements.txt
selenium>=4.27.1
requests>=2.32.3
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
